requirements.txt
openai==1.72.0
python-dotenv==1.0.0
streamlit==1.26.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
spacy==3.6.1
tiktoken==0.5.1
tqdm==4.65.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.0
faiss-cpu==1.7.4
langchain==0.0.267